// concierge/src/main/resources/projetoconcierge/src/app/models/aluno.ts
export class Aluno {
    id: number;
    cpf: number;
    senha: number;
    nome: string;
    telefone: number;
    data_nascimento: Date;
    sexo: string;
    num_rg: number;
    orgao_expedidor: string;
    naturalidade: string;
    nome_mae: string;
    nome_pai: string;
    estado_civil: string;
    rua: string;
    bairro: string;
    complemento: string;
    cidade: string;
    ocupacao: string;
    renda: number;
    escolaridade: string;
    tipo_residencia: string;
    pcdef: boolean;
    disp_horario: string;
    id_usuario:number;
  }
  